Add check on role before list modification

+++ lib/Alchemy/Phrasea/Controller/Prod/UsrLists.php
@@ -279,6 +279,11 @@ class UsrLists implements ControllerProviderInterface
 
           $list = $repository->findUserListByUserAndId($user, $list_id);
 
+          if($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_EDITOR)
+          {
+            throw new ControllerException(_('You are not authorized to do this'));
+          }
+
           $list->setName($list_name);
 
           $em->merge($list);
@@ -321,6 +326,11 @@ class UsrLists implements ControllerProviderInterface
           $user = $app['Core']->getAuthenticatedUser();
 
           $list = $repository->findUserListByUserAndId($user, $list_id);
+
+          if($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_ADMIN)
+          {
+            throw new ControllerException(_('You are not authorized to do this'));
+          }
 
           $em->remove($list);
           $em->flush();
@@ -369,6 +379,11 @@ class UsrLists implements ControllerProviderInterface
           $list = $repository->findUserListByUserAndId($user, $list_id);
           /* @var $list \Entities\UsrList */
 
+          if($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_EDITOR)
+          {
+            throw new ControllerException(_('You are not authorized to do this'));
+          }
+
           $entry_repository = $em->getRepository('\Entities\UsrListEntry');
 
           $user_entry = $entry_repository->findEntryByListAndUsrId($list, $usr_id);
@@ -411,8 +426,6 @@ class UsrLists implements ControllerProviderInterface
         $em   = $app['Core']->getEntityManager();
         $user = $app['Core']->getAuthenticatedUser();
 
-
-
         try
         {
           if (!is_array($request->get('usr_ids')))
@@ -425,7 +438,13 @@ class UsrLists implements ControllerProviderInterface
           $list = $repository->findUserListByUserAndId($user, $list_id);
           /* @var $list \Entities\UsrList */
 
+          if($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_EDITOR)
+          {
+            throw new ControllerException(_('You are not authorized to do this'));
+          }
+
           $inserted_usr_ids = array();
+
           foreach ($request->get('usr_ids') as $usr_id)
           {
             $user_entry = \User_Adapter::getInstance($usr_id, \appbox::get_instance());
@@ -503,7 +522,7 @@ class UsrLists implements ControllerProviderInterface
           if ($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_EDITOR)
           {
             $list = null;
-            throw new \Exception('You are not authorized to do this');
+            throw new \Exception(_('You are not authorized to do this'));
           }
 
         }
@@ -543,7 +562,7 @@ class UsrLists implements ControllerProviderInterface
 
           if ($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_EDITOR)
           {
-            throw new \Exception('You are not authorized to do this');
+            throw new \Exception(_('You are not authorized to do this'));
           }
 
           $new_owner = \User_Adapter::getInstance($usr_id, \appbox::get_instance());
@@ -614,7 +633,7 @@ class UsrLists implements ControllerProviderInterface
 
           if ($list->getOwner($user)->getRole() < \Entities\UsrListOwner::ROLE_ADMIN)
           {
-            throw new \Exception('You are not authorized to do this');
+            throw new \Exception(_('You are not authorized to do this'));
           }
 
           $owners_repository = $em->getRepository('\Entities\UsrListOwner');
